move remaining items
